task/support-cleanup * Move helper functions into spinner module
