+++ packages/pipeline-graph/src/main/support/SvgSpinner.tsx
@@ -1,7 +1,26 @@
 import * as React from 'react';
 
-import { describeArcAsPath } from './SVG';
 import { nodeStrokeWidth } from './StatusIcons';
+
+function polarToCartesian(centerX: number, centerY: number, radius: number, angleInDegrees: number) {
+    const angleInRadians = ((angleInDegrees - 90) * Math.PI) / 180.0;
+
+    return {
+        x: centerX + radius * Math.cos(angleInRadians),
+        y: centerY + radius * Math.sin(angleInRadians),
+    };
+}
+
+function describeArcAsPath(x: number, y: number, radius: number, startAngle: number, endAngle: number) {
+    const start = polarToCartesian(x, y, radius, endAngle);
+    const end = polarToCartesian(x, y, radius, startAngle);
+
+    const arcSweep = endAngle - startAngle <= 180 ? '0' : '1';
+
+    const d = ['M', start.x, start.y, 'A', radius, radius, 0, arcSweep, 0, end.x, end.y].join(' ');
+
+    return d;
+}
 
 interface Props {
     percentage: number;
